+++ src/model/Ball.hpp
@@ -3,20 +3,20 @@
 
 struct BallPositions
 {
-    int x = SCREEN_WIDTH/2;
-    int y = BALL_BOTTOM_POSITION;
+    float x = SCREEN_WIDTH/2;
+    float y = BALL_BOTTOM_POSITION;
 };
 
 struct BallDirection
 {
-    int x = 0;
-    int y = -BALL_DEFAULT_SPEED;
+    float x = 0;
+    float y = -BALL_DEFAULT_SPEED;
 };
 
 struct BallParamaters
 {
-    int radius = BALL_DEFAULT_RADIUS;
-    int speed = BALL_DEFAULT_SPEED;
+    float radius = BALL_DEFAULT_RADIUS;
+    float speed = BALL_DEFAULT_SPEED;
 };
 
 class Ball
@@ -26,10 +26,10 @@ public:
     void update();
     bool isLost() const;
     void reset();
-    void setPosition(int x, int y = BALL_BOTTOM_POSITION);
-    void setDirection(int dx, int dy);
-    void setRadius(int radius);
-    void setBallSpeed(int speed);
+    void setPosition(float x, float y = BALL_BOTTOM_POSITION);
+    void setDirection(float dx, float dy);
+    void setRadius(float radius);
+    void setBallSpeed(float speed);
     const BallPositions &getPositions() const;
     const BallDirection &getDirection() const;
     const BallParamaters &getParameters() const;
